person_personlike_detector/requirements.txt
ipython
matplotlib==3.10.3
numpy
opencv_python==4.11.0.86
pandas==2.3.0
Pillow==11.3.0
PyYAML==6.0.2
seaborn==0.13.2
streamlit==1.46.0
torch==2.7.1
torchvision==0.22.1
ultralytics==8.3.156
